Initial commit for 'msc_robot_arm' project
Started from instructions in the INSTALL.md of the
'Bare-Arduino-Project' Github repo and modifies for my taste in this
commit.
    - Not using a 'lib' folder, all code will be in the src/ dir
    - Removed a bunch of extra files that I won't need
    - Updated README for my project

Updated the Makefile and confirmed its functionality with the board, the
FooProject.cpp is the only code right now, and it just has some prints.

All previous commits in this repo belong to the 'Bare-Arduino-Project'