Adds servo_testing.cpp for testing arm motors
Code started from an example found online from the linxmotion forums
about inverse kinematics for the arm.  The IK function exists here but
is commented out (and still not tested).  This code is set up to test
individual AL5D servos via serial.  The loop requests a servo input and
a character command to 'go' via serial, and sets the motor to the given
input if it is valid.  The motor is set to the Gripper right now, but it
is a one line change to update it to one of the other 4 motors.  All are
tested and functional.