Fixes Elbow and Wrist home servo settings
The first measurement for the home position was taken so that the ulna
was parallel to the ground.  However, the elbow and wrist joints were
not exactly aligned with the ulna.  Since the joints are what really
matters, these new settings align the elbow-to-wrist joint imaginary
line with the ground, and the wrist-to-end-of-gripper imaginary line
with the ground.