Creates arm_controller derived class
This class will be responsible for all application specific control of
the Al5d robot arm, and as such, is a derived class of the Al5d.

This adds an override to the SetArm command that calculates the new
desired height using the error correction functions, and then delegates
back to the Al5d SetArm function to move to the new position.