Arm controller Update with speed governing
The SetArm function now only saves the next desired positions of the
arm.

A new Update function sets the arm's position, with a check to govern
the speed.  This check is based on a constant value (25mm) set as the
maximum movement per second for each axis.  The Update function takes in
the time since the last Update call to use for calculating the real-time
speed.

Because the ArmController object now must keep track of the current vs
desired positions, the object now also overloads the SetHomePosition
function.  This merely delegates to the Al5d and then uses the Getters
to save the current x, y, and z coordinates.