Printing induction sensor as arm height is lowered
This is a simple test all run through the main loop.  The arm and metal
detector are updated at given intervals, as well as the test output.
Each time the test updates (constant set at 2 seconds), it prints the
metal detector inductance value at the current height, and then
decreases the height by the defined increment (0.5 mm).  This continues
to run from the defined start point until the defined minimum height is
reached.