MetalScanner class created for defined area scanning
The MetalScanner object executes a state machine via its Update function
designed to scan an area defined by the given origin and desired
horizontal and depth travel bounds.  It first moves to a distance above
the origin, and desends straight downward until the desired inducance
delta is reached.  It then scans along the x and y axes by the defined
amounts, correcting the height based on the desired inductance along the
entire route.

Iniitialization of the object is done through a HardwareSetup function
to connect all hardware, and a SetReferenceInductance function to cache
the reference inducatnce value.

Setters are included for all parameters, as well as an isScanComplete
function and a ResetScan function so that it can be automatically
repeated.

Link to lucidchart state machine.
https://www.lucidchart.com/documents/edit/8c9fd1a9-f56d-4d4b-9284-004321b2eb55#