Adds Getters for current x, y, z in ArmController
While this may seem redundant since they also exist in the Al5d, the
height is different in these two scopes because of the correction
present in the ArmController.  This getter will return the desired
height, while the Al5d getter will return the real set value which is
effected by the height correction.  The x and y will be the same in both
scopes, but are added here as well for consistency and extensibility.