Adds defined area scanning state machine to MetalScanner
The Update function of the MetalScanner now executes a state machine
designed to scan an area defined by the given origin and desired
horizontal and depth travel bounds.  It first moves to a distance above
the origin, and desends straight downward until the desired inducance
delta is reached.  It then scans moves along the x and y axes by the
defined amounts, correcting the height based on the desired inductance
along the entire route.

Also includes isScanComplete and ResetScan functions so that the process
can be automatically repeated.

Link to lucidchart state machine.
https://www.lucidchart.com/documents/edit/8c9fd1a9-f56d-4d4b-9284-004321b2eb55#