Removes Gripper servo from Al5d
Because of the piece being used to to fix the sensor at the end of the
gripper, it must be sure to never move so that the servo does not get
damaged.  The gripper was never actually being controlled here anyway,
but now it is not even "Attach"ed so that the motor never runs at all.