Creates and implements ShapeMappingScanner
The ShapeMappingScanner must be given an origin, and then moves down to
the desired inductance and scans for edges in all four directions.  The
number of points for each edge is also settable.  It moves to an error
state if the origin is invalid, or if it moves out of bounds before
finding an edge in any direction.