WIP - shape mapping from origin in all 4 directions
The shape mapper now starts from the given origin, and scans starting in
the positive x direction until reaching an edge.  It then returns to the
origin and repeats for all four directions.

At this point, it just prints out the coordinates found for each edge.

This currently requires the user-defined origin to be above the sample,
and the sample edges to all be within the bounds of the arm movement.