Adds additional sampling to SetReferenceInductance
In the MetalScanner object, rather than updating the MetalDetector once
and using the first reading, this now takes 20 readings with 50ms delays
between, and then averages the two median values to use as reference.
This will eliminate the possibility of a noisy reading ruining the
reference inductance for an entire test.