WIP - adds arm controller error handling
The ArmController::Update return value is now read into a local
variable, which is checked during states that are moving the arm.  If it
receives and error when trying to set up for the scan, it moves to an
unrecoverable error state, as the origin and/or sample position must be
invalid.  If it receives an error when attempting a scan, it records the
edge as out of bounds (-1) for that direction, and then proceeds through
the rest of the scan.